Extract QuotableArgumentsImpl
