+++ src/org/elixir_lang/psi/QuotableArguments.java
@@ -7,5 +7,5 @@ import com.intellij.psi.PsiElement;
  * Created by luke.imhoff on 3/14/15.
  */
 public interface QuotableArguments extends PsiElement {
-    public OtpErlangObject[] quoteArguments();
+    OtpErlangObject[] quoteArguments();
 }
